refactor(pricing): move the pricing cards out of dashboard/
Root cause: the two components that ARE the public pricing page lived under
components/dashboard/billing/, because the dashboard's upgrade route was the
second consumer and got to name them. Ownership should follow the primary
surface, and the location is the thing that decides who may change them: while
they sat under dashboard/ any pricing redesign was a tier-2 change to a
dashboard directory rather than a tier-3 change to a marketing one.

PricingCardsSection and FeatureCompareGrid move to components/pricing/ with
their own barrel. billing-upgrade.tsx keeps using them, and says so in the new
barrel's comment so the borrowing is not rediscovered later as an accident.

No markup changed. Git detects both as renames with 2 and 5 lines differing,
which are the relative import paths losing a directory level. Verified /pricing
renders identically to before the move and /dashboard/billing/upgrade still
returns 200 with its imports resolving.

This exists so the pricing redesign that follows is a change to marketing code.
