+++ apps/vectreal-platform/app/components/pricing/index.ts
@@ -1,0 +1,11 @@
+/*
+  The public pricing surface. These two components render the pricing page and
+  are borrowed by the dashboard's billing upgrade route, which is why they are
+  not under `dashboard/`: the marketing page is the primary consumer and the one
+  whose design owns them.
+*/
+export { FeatureCompareGrid } from './feature-compare-grid'
+export {
+	PricingCardsSection,
+	type PricingCardsSectionProps
+} from './pricing-cards-section'
